# specs/001-multitenant-foundation/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Adsup Multi-tenant Foundation API
  version: 1.0.0
  description: Stable Module 1 contract. All tenant paths re-authorize active membership and scope.
servers:
  - url: http://localhost:3000
...
paths:
...
  /v1/tenants:
    post:
      tags: [Tenants]
      operationId: createTenant
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [name, timezone]
              properties:
                name: { type: string, minLength: 2, maxLength: 160 }
                timezone: { type: string, minLength: 3, maxLength: 64 }
      responses:
        '201':
          description: Tenant and owner membership created atomically
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TenantMembershipSummary' }
        '422': { $ref: '#/components/responses/ValidationProblem' }
        '409': { $ref: '#/components/responses/ConflictProblem' }
  /v1/invitations/accept:
    post:
      tags: [Tenants]
      operationId: acceptInvitation
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [token]
              properties:
                token: { type: string, minLength: 20, maxLength: 512 }
      responses:
        '201':
          description: Membership created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TenantMembershipSummary' }
        '422': { $ref: '#/components/responses/ValidationProblem' }
        '409': { $ref: '#/components/responses/ConflictProblem' }
        '410': { $ref: '#/components/responses/GoneProblem' }
  /v1/tenants/{tenantId}:
    get:
      tags: [Tenants]
      operationId: getTenant
      parameters: [{$ref: '#/components/parameters/TenantId'}]
      responses:
        '200':
          description: Tenant details
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Tenant' }
        '404': { $ref: '#/components/responses/NotFoundProblem' }
    patch:
      tags: [Tenants]
      operationId: updateTenantStatus
      x-permission: tenant.manage
      parameters:
        - $ref: '#/components/parameters/TenantId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateTenantStatus' }
      responses:
        '200':
          description: Tenant lifecycle state updated with audit
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Tenant' }
        '409': { $ref: '#/components/responses/ConflictProblem' }
  /v1/tenants/{tenantId}/invitations:
    post:
      tags: [Tenants]
      operationId: createInvitation
      x-permission: member.invite
      parameters:
        - $ref: '#/components/parameters/TenantId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateInvitation' }
      responses:
        '201':
          description: Plain token is returned once
          content:
            application/json:
              schema: { $ref: '#/components/schemas/InvitationCreated' }
        '403': { $ref: '#/components/responses/ForbiddenProblem' }
        '409': { $ref: '#/components/responses/ConflictProblem' }
    get:
      tags: [Tenants]
      operationId: listInvitations
      x-permission: member.invite
      parameters: [{$ref: '#/components/parameters/TenantId'}]
      responses:
        '200':
          description: Invitations without token hashes
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Invitation' }
  /v1/tenants/{tenantId}/invitations/{invitationId}/revoke:
    post:
      tags: [Tenants]
      operationId: revokeInvitation
      x-permission: member.invite
      parameters:
        - $ref: '#/components/parameters/TenantId'
        - $ref: '#/components/parameters/InvitationId'
        - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '200':
          description: Revoked or existing idempotent result
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Invitation' }
        '404': { $ref: '#/components/responses/NotFoundProblem' }
...
  /v1/tenants/{tenantId}/memberships:
    get:
      tags: [RBAC]
      operationId: listMemberships
      x-permission: member.read
      parameters:
        - $ref: '#/components/parameters/TenantId'
        - $ref: '#/components/parameters/Cursor'
      responses:
        '200':
          description: Cursor-paginated memberships
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MembershipPage' }
  /v1/tenants/{tenantId}/memberships/{membershipId}:
    patch:
      tags: [RBAC]
      operationId: updateMembership
      x-permission: member.manage
      parameters:
        - $ref: '#/components/parameters/TenantId'
        - $ref: '#/components/parameters/MembershipId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [reason]
              properties:
                displayName: { type: string, minLength: 2, maxLength: 120 }
                employeeCode: { type: [string, 'null'], maxLength: 40 }
                status: { enum: [ACTIVE, SUSPENDED, LEFT] }
                reason: { type: string, minLength: 3, maxLength: 500 }
      responses:
        '200':
          description: Membership updated with audit
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Membership' }
        '409': { $ref: '#/components/responses/ConflictProblem' }
  /v1/tenants/{tenantId}/memberships/{membershipId}/assignments:
    get:
      tags: [Organization]
      operationId: listMembershipAssignments
      x-permission: member.read
      parameters:
        - $ref: '#/components/parameters/TenantId'
        - $ref: '#/components/parameters/MembershipId'
        - name: at
          in: query
          required: false
          schema: { $ref: '#/components/schemas/Timestamp' }
        - name: includeHistory
          in: query
          required: false
          schema: { type: boolean, default: false }
      responses:
        '200':
          description: Effective or historical assignments
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Assignment' }
    post:
      tags: [Organization]
      operationId: createAssignment
      x-permission: member.manage
      parameters:
        - $ref: '#/components/parameters/TenantId'
        - $ref: '#/components/parameters/MembershipId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateAssignment' }
      responses:
        '201':
          description: Effective assignment created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Assignment' }
        '409': { $ref: '#/components/responses/ConflictProblem' }
  /v1/tenants/{tenantId}/roles:
    get:
      tags: [RBAC]
      operationId: listRoles
      x-permission: role.read
      parameters: [{$ref: '#/components/parameters/TenantId'}]
      responses:
        '200':
          description: Roles and permission codes
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Role' }
    post:
      tags: [RBAC]
      operationId: createRole
      x-permission: role.manage
      parameters:
        - $ref: '#/components/parameters/TenantId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateRole' }
      responses:
        '201':
          description: Custom role created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Role' }
        '409': { $ref: '#/components/responses/ConflictProblem' }
  /v1/tenants/{tenantId}/memberships/{membershipId}/role-bindings:
    post:
      tags: [RBAC]
      operationId: grantRoleBinding
      x-permission: role.manage
      parameters:
        - $ref: '#/components/parameters/TenantId'
        - $ref: '#/components/parameters/MembershipId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateRoleBinding' }
      responses:
        '201':
          description: Role binding granted
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RoleBinding' }
        '409': { $ref: '#/components/responses/ConflictProblem' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    TenantId:
      name: tenantId
      in: path
      required: true
      schema: { $ref: '#/components/schemas/Uuid' }
    MembershipId:
      name: membershipId
      in: path
      required: true
      schema: { $ref: '#/components/schemas/Uuid' }
    InvitationId:
      name: invitationId
      in: path
      required: true
      schema: { $ref: '#/components/schemas/Uuid' }
...
    Cursor:
      name: cursor
      in: query
      required: false
      schema: { type: string, maxLength: 512 }
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema: { type: string, minLength: 8, maxLength: 128 }
  responses:
    ValidationProblem:
      description: Invalid input
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
...
    ForbiddenProblem:
      description: Actor lacks permission/scope
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    NotFoundProblem:
      description: Resource absent or intentionally hidden outside scope
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    ConflictProblem:
      description: State, concurrency, last-owner or idempotency fingerprint conflict
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    GoneProblem:
      description: Invite expired, revoked or exhausted
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
...
  schemas:
    Uuid:
      type: string
      format: uuid
    Timestamp:
      type: string
      format: date-time
...
    Tenant:
      type: object
      required: [id, name, slug, timezone, status]
      properties:
        id: { $ref: '#/components/schemas/Uuid' }
        name: { type: string }
        slug: { type: string }
        timezone: { type: string }
        status: { enum: [ACTIVE, SUSPENDED, CLOSED] }
    UpdateTenantStatus:
      type: object
      additionalProperties: false
      required: [status, reason]
      properties:
        status: { enum: [ACTIVE, SUSPENDED, CLOSED] }
        reason: { type: string, minLength: 3, maxLength: 500 }
    TenantMembershipSummary:
      type: object
      required: [tenant, membership]
      properties:
        tenant: { $ref: '#/components/schemas/Tenant' }
        membership: { $ref: '#/components/schemas/Membership' }
    Membership:
      type: object
      required: [id, tenantId, userId, displayName, status, version]
      properties:
        id: { $ref: '#/components/schemas/Uuid' }
        tenantId: { $ref: '#/components/schemas/Uuid' }
        userId: { $ref: '#/components/schemas/Uuid' }
        displayName: { type: string }
        employeeCode: { type: [string, 'null'] }
        status: { enum: [INVITED, ACTIVE, SUSPENDED, LEFT] }
        version: { type: integer, minimum: 1 }
    MembershipPage:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/Membership' }
        nextCursor: { type: [string, 'null'] }
...
    CreateInvitation:
      type: object
      additionalProperties: false
      required: [type, roleId, expiresAt]
      properties:
        type: { enum: [DIRECT, GROUP_LINK] }
        roleId: { $ref: '#/components/schemas/Uuid' }
        branchId: { anyOf: [{$ref: '#/components/schemas/Uuid'}, {type: 'null'}] }
        maxUses: { type: integer, minimum: 1, maximum: 1000, default: 1 }
        expiresAt: { $ref: '#/components/schemas/Timestamp' }
    Invitation:
      type: object
      required: [id, type, state, maxUses, useCount, expiresAt]
      properties:
        id: { $ref: '#/components/schemas/Uuid' }
        type: { enum: [DIRECT, GROUP_LINK] }
        state: { enum: [ACTIVE, EXPIRED, REVOKED, EXHAUSTED] }
        tokenHint: { type: string }
        maxUses: { type: integer }
        useCount: { type: integer }
        expiresAt: { $ref: '#/components/schemas/Timestamp' }
    InvitationCreated:
      allOf:
        - $ref: '#/components/schemas/Invitation'
        - type: object
          required: [token, inviteUrl]
          properties:
            token: { type: string }
            inviteUrl: { type: string, format: uri }
    CreateAssignment:
      type: object
      additionalProperties: false
      required: [branchId, effectiveFrom, reason]
      properties:
        branchId: { $ref: '#/components/schemas/Uuid' }
        departmentId: { anyOf: [{$ref: '#/components/schemas/Uuid'}, {type: 'null'}] }
        positionId: { anyOf: [{$ref: '#/components/schemas/Uuid'}, {type: 'null'}] }
        effectiveFrom: { $ref: '#/components/schemas/Timestamp' }
        effectiveTo: { anyOf: [{$ref: '#/components/schemas/Timestamp'}, {type: 'null'}] }
        reason: { type: string, minLength: 3, maxLength: 500 }
    Assignment:
      allOf:
        - $ref: '#/components/schemas/CreateAssignment'
        - type: object
          required: [id, tenantId, membershipId]
          properties:
            id: { $ref: '#/components/schemas/Uuid' }
            tenantId: { $ref: '#/components/schemas/Uuid' }
            membershipId: { $ref: '#/components/schemas/Uuid' }
    CreateRole:
      type: object
      additionalProperties: false
      required: [code, name, permissionCodes, reason]
      properties:
        code: { type: string, pattern: '^[A-Z0-9_-]{2,40}$' }
        name: { type: string, minLength: 2, maxLength: 120 }
        permissionCodes:
          type: array
          uniqueItems: true
          items: { type: string, minLength: 2, maxLength: 80 }
        reason: { type: string, minLength: 3, maxLength: 500 }
    Role:
      type: object
      required: [id, code, name, kind, permissionCodes]
      properties:
        id: { $ref: '#/components/schemas/Uuid' }
        code: { type: string }
        name: { type: string }
        kind: { enum: [SYSTEM, CUSTOM] }
        permissionCodes:
          type: array
          items: { type: string }
    CreateRoleBinding:
      type: object
      additionalProperties: false
      required: [roleId, scopeType, effectiveFrom, reason]
      properties:
        roleId: { $ref: '#/components/schemas/Uuid' }
        scopeType: { enum: [TENANT, BRANCH] }
        branchId: { anyOf: [{$ref: '#/components/schemas/Uuid'}, {type: 'null'}] }
        effectiveFrom: { $ref: '#/components/schemas/Timestamp' }
        effectiveTo: { anyOf: [{$ref: '#/components/schemas/Timestamp'}, {type: 'null'}] }
        reason: { type: string, maxLength: 500 }
    RoleBinding:
      allOf:
        - $ref: '#/components/schemas/CreateRoleBinding'
        - type: object
          required: [id, membershipId]
          properties:
            id: { $ref: '#/components/schemas/Uuid' }
            membershipId: { $ref: '#/components/schemas/Uuid' }
...
    Problem:
      type: object
      additionalProperties: false
      required: [type, title, status, code, correlationId]
      properties:
        type: { type: string, format: uri-reference }
        title: { type: string }
        status: { type: integer, minimum: 400, maximum: 599 }
        code: { type: string, pattern: '^[A-Z0-9_]+$' }
        detail: { type: string }
        correlationId: { type: string }
        fieldErrors:
          type: array
          items:
            type: object
            required: [path, code, message]
            properties:
              path: { type: string }
              code: { type: string }
              message: { type: string }
security:
  - bearerAuth: []